Refactor JdbcLinkUpdate to LinkUpdaterImpl
