+++ JavaService/scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/jooq/JooqLinkService.java
@@ -36,7 +36,6 @@ import java.util.stream.Collectors;
 public class JooqLinkService implements LinkService {
 
     private static final int COUNT_LINK_LIMIT = 10;
-    private final DSLContext dsl;
     private final ClientManager clientManager;
 
     private final JooqRequestLinkRepository jooqRequestLink;
