+++ JavaService/scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/web/sheduler/LinkUpdaterScheduler.java
@@ -5,7 +5,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.scrapper.service.jdbc.JdbcLinkUpdater;
+import ru.tinkoff.edu.java.scrapper.service.LinkUpdaterImpl;
 
 @Component
 @RequiredArgsConstructor
@@ -13,11 +13,11 @@ public class LinkUpdaterScheduler {
     private static final Logger log =
             LoggerFactory.getLogger(LinkUpdaterScheduler.class);
 
-    private final JdbcLinkUpdater jdbcLinkUpdater;
+    private final LinkUpdaterImpl linkUpdaterImpl;
 
     @Scheduled(fixedDelayString = "#{schedulerIntervalMs}")
     public void update() {
-        int countUpdate = jdbcLinkUpdater.update();
+        int countUpdate = linkUpdaterImpl.update();
         log.info("Update info about urls. Count Link update: " + countUpdate);
     }
 }
